Move zmq_helpers only used in tests into tests to avoid coverage drop
