+++ flowapi/tests/unit/test_geography.py
@@ -5,7 +5,7 @@
 import pytest
 from json import loads
 
-from flowapi.zmq_helpers import ZMQReply
+from tests.unit.zmq_helpers import ZMQReply
 
 
 @pytest.mark.asyncio
